chore: better class methods on `AntReader`
