+++ modules/ant/src/heartrate.py
@@ -8,12 +8,14 @@ class HeartRate(AntDevice):
         super().__init__(node, sensor_id)
 
         self._device_type_id = device_type.value
-        self._last_data_time = None
+
+        # data store
+        self._heartrate = 0
 
         # inizializzazione del channel ant
         self._init_channel()
 
-    # NOTE: specializzazione metodi astratti di `AntDevice`
+    # Specializzazione metodi astratti di `AntDevice`
 
     def _init_channel(self):
         if self._channel is None:
@@ -27,32 +29,21 @@ class HeartRate(AntDevice):
         self._channel.set_search_timeout(255)
         self._channel.set_rf_freq(57)
 
-        # 69  -> ID DELLA MIA FASCIA CARDIO, METTERE 0 PER TUTTE LE FASCIE
-        # 120 -> DEVICE ID DEL SENSORE DELLA FC
-
-        # hr_sensor_id = self._settings.hr_sensor_id
         self._channel.set_id(self._sensor_id, self._device_type_id, 0)
 
         # open channel
         self._channel.open()
 
     def _receive_new_data(self, data):
-        self._heartrate = data[7]
-        self._last_data_time = self._current_time()
+        self._payload = data
+        self._last_data_read = self._current_time()
 
     def read_data(self) -> dict:
-        if not self._is_active():
-            self._heartrate = 0
+        self._heartrate = self._get_heartrate() if self._is_active() else 0
 
         return {"heartrate": self._heartrate}
 
-    # NOTE: metodi propri della classe
+    # Metodi propri della classe
 
-    def _current_time(self):
-        return time._unix_time()
-
-    def _elapsed_time(self):
-        return self._current_time() - self._last_data_time
-
-    def _is_active(self):
-        return self._last_data_time and self._elapsed_time() < 5
+    def _get_heartrate(self):
+        return self._payload[7]
